requirements.txt
httpx
beautifulsoup4
openai
pydantic
pandas
python-dotenv
openpyxl
playwright
